event-lite/backend/requirements.txt
Flask==3.0.3
flask-socketio==5.3.6
flask-cors==4.0.1
python-socketio==5.11.2
eventlet==0.36.1
requests==2.31.0
python-dotenv==1.0.0